requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
google-cloud-storage==2.10.0
Pillow==10.0.0
flask-login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
flask-talisman==1.0.0
gunicorn==21.2.0
numpy==2.2.3